+++ services/core/java/core-flow-service/src/main/java/org/ubiquia/core/flow/service/registrar/FlowMessageRegistrar.java
@@ -52,7 +52,7 @@ public class FlowMessageRegistrar {
         }
 
         var nodeEntity = nodeOpt.get();
-        var graphEntity = nodeEntity.getGraph();
+        var graphEntity = nodeEntity.getParentGraph();
 
         var flowEntity = new FlowEntity();
         flowEntity.setGraph(graphEntity);