+++ services/core/java/core-flow-service/src/test/java/org/ubiquia/core/flow/TestHelper.java
@@ -70,7 +70,7 @@ public class TestHelper {
 
         var nodeEntity = this
             .nodeRepository
-            .findByGraphNameAndName(graphName, nodeName)
+            .findByParentGraphNameAndName(graphName, nodeName)
             .get();
 
         var node = nodeMap